+++ storybook/src/templates.ts
@@ -70,6 +70,7 @@ import {
   Logo,
   MapControls,
   MapLayer,
+  MapMessage,
   MarkBar,
   Modal,
   Navbar,
@@ -183,6 +184,7 @@ import { coreListButton } from "./components/list-button/list-button.core-templa
 import { coreLogo } from "./components/logo/logo.core-template";
 import { coreMapControls } from "./components/map-controls/map-controls.core-template";
 import { coreMapLayer } from "./components/map-layer/map-layer.core-template";
+import { coreMapMessage } from "./components/map-message/map-message.core-template";
 import { coreMarkBar } from "./components/mark-bar/mark-bar.core-template";
 import { coreModal } from "./components/modal/modal.core-template";
 import { cssModal } from "./components/modal/modal.css-template";
@@ -216,7 +218,6 @@ import { coreToggletip } from "./components/toggletip/toggletip.core-template";
 import { coreTooltip } from "./components/tooltip/tooltip.core-template";
 import { coreTreeView } from "./components/tree-view/tree-view.core-template";
 import { coreViewerGrid } from "./components/viewer-grid/viewer-grid.core-template";
-
 // End: import templates
 
 export interface Components {
@@ -288,6 +289,7 @@ export interface Components {
   logo: Logo;
   mapControls: MapControls;
   mapLayer: MapLayer<TemplateResult>;
+  mapMessage: MapMessage;
   markBar: MarkBar;
   modal: Modal<TemplateResult>;
   navbar: Navbar<TemplateResult>;
@@ -364,6 +366,7 @@ templateContainer.add(coreButtonGroup);
 templateContainer.add(coreCard);
 templateContainer.add(coreCardContainer);
 templateContainer.add(coreContactInformation);
+templateContainer.add(coreCursorTooltip);
 templateContainer.add(coreDatePicker);
 templateContainer.add(coreDocumentCard);
 templateContainer.add(coreDocumentComponent);
@@ -384,6 +387,7 @@ templateContainer.add(coreListButton);
 templateContainer.add(coreLogo);
 templateContainer.add(coreMapControls);
 templateContainer.add(coreMapLayer);
+templateContainer.add(coreMapMessage);
 templateContainer.add(coreMarkBar);
 templateContainer.add(coreModal);
 templateContainer.add(coreOnboardingTip);
@@ -458,5 +462,4 @@ templateContainer.add(cssSearchBar);
 templateContainer.add(cssShoppingCart);
 templateContainer.add(cssTile);
 templateContainer.add(cssTileGrid);
-templateContainer.add(coreCursorTooltip);
 // End: add templates to container